Moving nvtx.h to common/cuda
